requirements.txt
colorlog==6.8.2
GitPython==3.1.43